+++ spring-cloud-tencent-commons/src/main/java/com/tencent/cloud/common/constant/ContextConstant.java
@@ -36,6 +36,11 @@ public final class ContextConstant {
 	 */
 	public static final String UTF_8 = StandardCharsets.UTF_8.name();
 
+	/**
+	 * Default registry heartbeat time interval, default: 5000 (ms).
+	 */
+	public static final Integer DEFAULT_REGISTRY_HEARTBEAT_TIME_INTERVAL = 5000;
+
 	private ContextConstant() {
 	}