+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - [Bugfix: optimize ratelimit actuator](https://github.com/Tencent/spring-cloud-tencent/pull/419)
 - [Feature: add rate limit filter debug log](https://github.com/Tencent/spring-cloud-tencent/pull/436)
 - [Optimize: add EncodeTransferMedataRestTemplateInterceptor to RestTemplate](https://github.com/Tencent/spring-cloud-tencent/pull/439)
+- [Add configurable heartbeat interval support](https://github.com/Tencent/spring-cloud-tencent/pull/443)